refactor(trusty-code): split CodeEngine's engine.rs under the 500-SLOC cap
engine.rs was 759 SLOC (issue #610's mechanical cap: 500 for production
files). Split EngineState (session/cache/streaming logic) into
engine_state.rs, the pure Event->ReplEvent mapping into session_events.rs,
and the workstream-activation SSE loop into workstream_subscription.rs.
Tests move to the sibling engine_tests.rs, matching this crate's existing
_tests.rs convention (serve/http.rs -> http_tests.rs). No behavior change;
verified via `scripts/check_line_cap.sh` and the full test suite.

Part of #3415
